requirements.txt
# Core Pipeline & IaC
boto3>=1.35.0
localstack>=4.0.0
aws-sam-translator
# AI & LLM Integration
openai>=1.0.0
openai-agents
pydantic>=2.0.0
python-dotenv
requests
mcp
agents
# Data Analysis & Research Results (src/validation/)
pandas
numpy
matplotlib
openpyxl
scipy
networkx
# Utilities
tqdm
pyyaml
rich
plantweb
typing-extensions
hcl2